backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
pycryptodome==3.19.0
numpy==1.26.2
python-dotenv==1.0.0
